Button controls now show and hide correctly
Before logging in, only Sign Up and Sign In appear. After login, Change
Password and Sign Out appear. For the new user who has not created a
Profile yet, the All Concerts and My Concerts buttons are disabled.
Only the Create Profile button is available. Once a Profile is created,
the user can see all concerts or her/his saved concerts. A Filter is
provided for searching the list of available concerts.