Save button in All Concerts is disabled
The Save button text is also changed to Saved. Now, concerts cannot
be saved more than once.

This required getting the user's saved concerts when the user logs in.
Prior to the All Concerts list or All Concerts Filtered list, a new
array is created with an additional boolean which indicates whether
or not the concert has been saved.

For debugging, the concert id is also in the list and must be removed
before submission.