All Concerts and My Concerts now enabled after new user creates profile
Typing into the About Me textarea in the profile dialog now starts at
the beginning of the control. For textarea control, there can be no
spaces opening and closing tags: <textarea class="about"...></textarea>.
The </textarea> was on the next line for readable creating spaces
and / or tabs in the control.